refactor: 컴포넌트 폴더 구조 변경 #11
- ResponsivePanel, ResponsiveWrapper, DrawerBottom, SheetSide 컴포넌트를 components/panel/로 이동
- TransactionSection을 components/transaction/으로 이동
