+++ fe/src/app/(app)/page.tsx
@@ -1,8 +1,8 @@
 import { Button } from '@/components/ui/button';
 import { DUMMY_TRANSACTIONS } from '@/constants/dummyData';
 import LogoutButton from '@/components/LogoutButton';
-import ResponsivePanel from '@/components/common/ResponsivePanel';
-import TransactionSection from '@/components/common/TransactionSection';
+import ResponsivePanel from '@/components/panel/ResponsivePanel';
+import TransactionSection from '@/components/transaction/TransactionSection';
 
 export default function Home() {
   return (
